Allow hyphenated db names in role definition
Currently we are using a regex to allow hyphens in DB names. Hyphens in DB names are fine with mongo, and we have existing DBs with hyphenated names.
However, in the roles array we are using a different regex for the DB part of the role which only allows \w. This should be changed to match the DB criteria.